Implement React Context
Stores data of the user that is logged in.

Includes a useEffect that runs when mounted - Checks if there is a user logged in & sets a loading state to prevent UI from flashing. Without the loading state, the Navbar would flash 'LOGIN / REGISTER' before showing 'MY ACCOUNT / LOGOUT' if the user were logged in